+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [UNRELEASED]
+### Changed
+- [LIBRARY] Keep original line number in the JsonML tree for HTML comments.
+
+
 ## [2.0.104] - 20200304
 ### Changed
 - [INTERNAL] More linting